% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_molecules.R
\name{add_molecules}
\alias{add_molecules}
\title{Adds Molecule Dataframes, Helper Function to get_peptide_coverage}
\usage{
add_molecules(..., CapNegatives = TRUE)
}
\arguments{
\item{...}{Any number of molecules from the make_molecule class}

\item{CapNegatives}{A TRUE/FALSE to indicate whether negative elements should be
capped at 0. Useful for modifications where elements are lost. Default is TRUE.}
}
\description{
This simple function allows for two molecule data frames from make_molecule to be added together
}
\details{
Returns a molecule dataframe which is the summation of the provided molecule dataframes
}
\examples{
\dontrun{

# Add two molecules together
add_molecules(
  make_molecule(AtomList = list("C" = 2, "H" = 3, "O" = 1, "P" = 0)),
  make_molecule(AtomList = list("C" = 1, "H" = 4, "O" = 0, "N" = 1)),
  make_molecule(AtomList = list("C" = 6))
)

}

}
