% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molform_functions.R
\name{add_molforms}
\alias{add_molforms}
\title{Add molecular formula objects}
\usage{
add_molforms(..., CapNegatives = TRUE)
}
\arguments{
\item{...}{Any number of molecules from the as.molform class}

\item{CapNegatives}{A TRUE/FALSE to indicate whether negative elements should be
capped at 0. Useful for modifications where elements are lost. Default is TRUE.}
}
\description{
Add molform objects together
}
\examples{
\dontrun{

add_molforms(as.molform("C6H12O6"), as.molform("H2Fr2Sr2Ar5"), as.molform("H-1C-2"))

}
}
