% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_spectrum_plot.R
\name{annotated_spectrum_plot}
\alias{annotated_spectrum_plot}
\title{Annotate a Spectrum with Fragments}
\usage{
annotated_spectrum_plot(
  PeakData,
  MatchedPeaks = NULL,
  IncludeIsotopes = TRUE,
  IncludeLabels = TRUE,
  LabelSize = 4,
  LabelDistance = 0.5,
  Interactive = FALSE
)
}
\arguments{
\item{PeakData}{Object of the peak_data class from get_peak_data.}

\item{MatchedPeaks}{Object of the matched_peaks class from get_matched_peaks. Default is NULL.}

\item{IncludeIsotopes}{A logical to indicate whether isotopes should be included in the plot. Default is TRUE.}

\item{IncludeLabels}{A logical to indicate whether labels should be included in the plot. Default is TRUE.}

\item{LabelSize}{A numeric indicating the size of the label in ggplot dimensions. Default is 4.}

\item{LabelDistance}{A numeric indicating the distance from each peak the label should be. Default is 0.5 M/Z.}

\item{Interactive}{A logical to determine whether the plot should be interactive or not. For plots
with more than 10,000 peaks, it is advantageous to set to FALSE. Default is TRUE.}
}
\description{
Generates an experimental spectrum with calculated peptide fragments.
   Plot can be interactive or not.
}
\examples{
\dontrun{

# Test bottom up data
BU_Peak <- get_peak_data(ScanMetadata = BU_ScanMetadata, ScanNumber = 31728)
BU_Match <- get_matched_peaks(ScanMetadata = BU_ScanMetadata, PeakData = BU_Peak)

annotated_spectrum_plot(PeakData = BU_Peak, Interactive = TRUE)
annotated_spectrum_plot(PeakData = BU_Peak, MatchedPeaks = BU_Match, IncludeLabels = FALSE, Interactive = TRUE)
annotated_spectrum_plot(PeakData = BU_Peak, MatchedPeaks = BU_Match, IncludeLabels = TRUE)
}

}
