% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molform_functions.R
\name{as.molform}
\alias{as.molform}
\title{Make a molecular formula object}
\usage{
as.molform(MolForm)
}
\arguments{
\item{MolForm}{A molecular formula written as a string like "C6H12O6". Only 
elements up to uranium are supported.}
}
\description{
This function generates a molecular formula object.
}
\details{
A list of length 92 (Hydrogen to Uranium) is stored in the attributes 
   and used for quick mathematics. All numerics are rounded to the nearest whole
   number.
}
\examples{
\dontrun{

# Generate a molecular formula object
as.molform("C6H12O6")

# This should fail, as 'Ot' and 'Lp' are not elements 
as.molform("C6H12Ot5Lp2")

}

}
