% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molform_functions.R
\name{calcIsoProfile}
\alias{calcIsoProfile}
\title{Calculate an isotope profile using isopat}
\usage{
calcIsoProfile(molform, min_abundance = 0.1)
}
\arguments{
\item{molform}{An object of the as.molform class}

\item{min_abundance}{Minimum abundance for calculating isotopes. Default is 0.1.}
}
\description{
Generates an isotope profile using [isopat](https://github.com/cran/isopat)
}
\examples{
\dontrun{

calcIsoProfile(molform = as.molform("C6H12O6"), min_abundance = 0.1)

}
}
