% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molform_functions.R
\name{collapse_molform}
\alias{collapse_molform}
\title{Collapse a molform object to a simple molecular formula}
\usage{
collapse_molform(molform)
}
\description{
A simple printing option for an object of the as.molform class.
}
\examples{
\dontrun{
collapse_molform(as.molform("H12C6O6"))
}
}
