% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molform_functions.R
\name{convert_proforma}
\alias{convert_proforma}
\title{Convert a proforma string to a PTM object}
\usage{
convert_proforma(proforma)
}
\arguments{
\item{proforma}{A string written in the format "M.S[Methyl]S[22]S[23].V"}
}
\description{
Converts proforma strings to a PTM object that get_matched_peaks uses
    to calculate fragmentation patterns. If there is no modification annotation,
    the cleaned sequence is returned.
}
\examples{
\dontrun{

convert_proforma("M.(S)[Acetyl]ATNNIAQARKLVEQLRIEAGIERIKVSKAASDLMSYCEQHARNDPLLVGVPASENPFKDK(KPCIIL)[-52.9879].")

convert_proforma("M.SS[Methyl]S.V")

convert_proforma("M.S[Methyl]S[22]S[23].V")

convert_proforma("TESTTESTTEST")

}
}
\author{
Degnan, David. Flores, Javier.
}
