% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ion_annotations.R
\name{count_ion_annotations}
\alias{count_ion_annotations}
\title{Generates a Summary Table of the Number of Ions per Residue}
\usage{
count_ion_annotations(MatchedPeaks, IncludeIsotopes = FALSE)
}
\arguments{
\item{MatchedPeaks}{A matched_peaks object generated by get_matched peaks. Required.}

\item{IncludeIsotopes}{A logical to indicate whether isotopes should be included. Default is FALSE.}
}
\description{
A simple summary table of the number of ions per residues
}
\examples{
\dontrun{

# Test bottom up data
BU_Peak <- get_peak_data(ScanMetadata = BU_ScanMetadata, ScanNumber = 31728)
BU_Match <- get_matched_peaks(ScanMetadata = BU_ScanMetadata, PeakData = BU_Peak)

count_ion_annotations(MatchedPeaks = BU_Match)

}

}
