% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_bar_plot.R
\name{coverage_bar_plot}
\alias{coverage_bar_plot}
\title{Generates a bar plot of identified residues for a specific protein}
\usage{
coverage_bar_plot(PeptideCoverage, Interactive = FALSE)
}
\arguments{
\item{PeptideCoverage}{A peptide_coverage object generate by get_peptide_coverage. Required.}

\item{Interactive}{A TRUE/FALSE to indicate whether the plot should be interactive or not. Default is FALSE.}
}
\description{
Generates a bar plot of counts of identified residues across fragments
    for a specific protein
}
\examples{
\dontrun{

# Make Peptide Coverage Object
PeptideCoverage <- get_peptide_coverage(ScanMetadata = BU_ScanMetadata, ProteinTable = ProteinTable, ProteinID = "SO_0225")

# Generate peptide coverage plot
coverage_bar_plot(PeptideCoverage)
coverage_bar_plot(PeptideCoverage, Interactive = TRUE)

}
}
