% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_lit_seq_plot.R
\name{coverage_lit_seq_plot}
\alias{coverage_lit_seq_plot}
\title{Displays the literature sequence with identified residues for a specific protein}
\usage{
coverage_lit_seq_plot(PeptideCoverage)
}
\arguments{
\item{PeptideCoverage}{A "peptide_coverage" object generated by get_peptide_coverage. Required.}
}
\description{
A plot of the literature sequence with identified fragment residues in green
}
\examples{
\dontrun{

# Make Peptide Coverage Object
PeptideCoverage <- get_peptide_coverage(ScanMetadata = BU_ScanMetadata, ProteinTable = ProteinTable, ProteinID = "SO_0225")

# Make plot
coverage_lit_seq_plot(PeptideCoverage)

}

}
