% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plot.R
\name{coverage_plot}
\alias{coverage_plot}
\title{Generate a plot of peptide coverage for a specific protein}
\usage{
coverage_plot(PeptideCoverage, ColorByScore = NULL, Interactive = FALSE)
}
\arguments{
\item{PeptideCoverage}{A peptide_coverage object generate by get_peptide_coverage. Required.}

\item{ColorByScore}{A character to indicate whether the coverage plot should
should be colored by QValue, Score, or not at all. Acceptable values are "QValue",
"Score", and NULL. Default is NULL.}

\item{Interactive}{A TRUE/FALSE to indicate whether the plot should be interactive or not. Default is FALSE.}
}
\description{
A plot which shows where identified peptides map to the literature sequence
}
\examples{
\dontrun{

# Make Peptide Coverage Object
PeptideCoverage <- get_peptide_coverage(ScanMetadata = BU_ScanMetadata, ProteinTable = ProteinTable, ProteinID = "SO_0225")

# Generate peptide coverage plot
coverage_plot(PeptideCoverage)
coverage_plot(PeptideCoverage, ColorByScore = "Score", Interactive = TRUE)

}

}
