% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_heatmap_plot.R
\name{error_heatmap_plot}
\alias{error_heatmap_plot}
\title{Generate a Heatmap of PPM Errors per Ion Type per Residue}
\usage{
error_heatmap_plot(MatchedPeaks, IncludeIsotopes = FALSE, Interactive = TRUE)
}
\arguments{
\item{MatchedPeaks}{A matched_peaks object generated by get_matched peaks. Required.}

\item{IncludeIsotopes}{A logical to indicate whether isotopes should be included. Default is FALSE
for cleaner heatmaps.}

\item{Interactive}{A logical to determine whether the plot should be interactive or not. Default is TRUE.}
}
\description{
Make a Heatmap where x is the peptide residue (in order from N to C-terminus), y is the ion type, and z is the ppm error
}
\examples{
\dontrun{

# Test bottom up data
BU_Peak <- get_peak_data(ScanMetadata = BU_ScanMetadata, ScanNumber = 31728)
BU_Match <- get_matched_peaks(ScanMetadata = BU_ScanMetadata, PeakData = BU_Peak)

error_heatmap_plot(MatchedPeaks = BU_Match, IncludeIsotopes = FALSE)
error_heatmap_plot(MatchedPeaks = BU_Match, IncludeIsotopes = TRUE)
error_heatmap_plot(MatchedPeaks = BU_Match, IncludeIsotopes = TRUE, Interactive = FALSE)

}

}
