% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molform_functions.R
\name{get_aa_molform}
\alias{get_aa_molform}
\title{Calculate a molecular formula for an amino acid sequence}
\usage{
get_aa_molform(sequence)
}
\arguments{
\item{sequence}{The amino acid sequence.}
}
\description{
Calculate a molecular formula for an amino acid sequence
}
\examples{
\dontrun{

getAAMolForm("TESTTESTTESTTEST")

}
}
