% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molform_functions.R
\name{get_monoisotopic}
\alias{get_monoisotopic}
\title{Get monoisotopic mass from the molecular formula}
\usage{
get_monoisotopic(molform)
}
\arguments{
\item{molform}{An object of the as.molform class}
}
\description{
Calculate the monoisotopic mass for a molform object. No charge is assumed.
}
\examples{
\dontrun{

get_monoisotopic(as.molform("C6H12O6"))

}
}
