% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ms1ft.R
\name{get_ms1ft}
\alias{get_ms1ft}
\title{Generates the "ms1ft" object, which is Top-Down MS1 Feature data}
\usage{
get_ms1ft(MS1FTPath, TargetsPath = NULL, TrimColumns = TRUE)
}
\arguments{
\item{MS1FTPath}{Path to the MS1FT txt file, which is a tsv. Required.}

\item{TargetsPath}{Optional path to an IC targets file which links MS1FT features to proteins. Default is NULL.}

\item{TrimColumns}{A logical to indicate whether all columns or only columns relevant to
promex_feature_plot should be returned. Default is TRUE.}
}
\description{
Reads in ms1ft from the ProMex algorithm in InformedProteomics, and
   associates features to protein data from the Ic_Targets file from MSPathFinderT.
   All of this data is used to generate the promex feature plot.
}
\details{
The data.table outputted by this function contains many columns. Listed below are the main ones.
\tabular{ll}{
FeatureID \tab An identified mass across elution (retention) times is given a feature ID, typically a single number. \cr
\tab \cr
MinElutionTime \tab The smallest retention time of the feature \cr
\tab \cr
MaxElutionTime \tab The largest retention time of the feature \cr
\tab \cr
MonoMass \tab The monoisotopic mass of the feature. \cr
\tab \cr
Abundance \tab The relative abundance of the feature. \cr
\tab \cr
ProteinName \tab If the Targets file is provided, the name of the protein associated to that feature. \cr
\tab \cr
}
}
\examples{
\dontrun{

# Test top down data
tmpdir <- tempdir()
download.file("https://raw.githubusercontent.com/EMSL-Computing/PSpecteR/master/pspecter_container/TestFiles/TopDown/TopDown.ms1ft",
              file.path(tmpdir, "TopDown.ms1ft"))
download.file("https://raw.githubusercontent.com/EMSL-Computing/PSpecteR/master/pspecter_container/TestFiles/TopDown/TopDown_IcTarget.tsv",
              file.path(tmpdir, "TopDown_IcTarget.tsv"))

# Generate object
MS1FT <- get_ms1ft(MS1FTPath = file.path(tmpdir, "TopDown.ms1ft"),
                   TargetsPath = file.path(tmpdir, "TopDown_IcTarget.tsv"))

}

}
