% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molform_functions.R
\name{get_mw}
\alias{get_mw}
\title{Get average mass from the molecular formula}
\usage{
get_mw(molform)
}
\arguments{
\item{molform}{An object of the as.molform class}
}
\description{
Calculate the average mass for a molform object.
}
\examples{
\dontrun{

get_mw(as.molform("C6H12O6"))

}
}
