% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peak_data.R
\name{get_peak_data}
\alias{get_peak_data}
\title{Extracts peak data from a raw or mzML file}
\usage{
get_peak_data(ScanMetadata, ScanNumber, MinAbundance = 0)
}
\arguments{
\item{ScanMetadata}{Object of the scan_metadata class from get_scan_metadata. Required.}

\item{ScanNumber}{Integer indicating which scan number to pull the peak data. Required.}

\item{MinAbundance}{Filter out peaks with an abundance below threshold. Ranges from 0-100. Default is 0.}
}
\description{
Extract the M/Z and Intensity data for a specific scan number from
   a raw or mzML file. Use "make_peak_data" to create a "peak_data" object.
}
\details{
The data.table outputted by this function contains both the M/Z and Intensity vectors of the spectra.

Information regarding user input and calculated metrics are stored in the attributes
\tabular{ll}{
ScanNumber \tab The scan number of the spectra, as given by the user in the get_peak_data parameters \cr
\tab \cr
TotalNumberPeaks \tab The total number of peaks stored in the MS file. \cr
\tab \cr
NumberPeaksPostFilter \tab The number of peaks remaining post filter. \cr
\tab \cr
PercentagePeaksRemain \tab The percentage or proportion of peaks remaining post-filter. This is NumberPeaksPostFilter/TotalNumberPeaks. \cr
\tab \cr
MSPath \tab The path to the MS file \cr
}
}
\examples{
\dontrun{

# Pull peak data from various examples
peak1 <- get_peak_data(ScanMetadata = BU_ScanMetadata, ScanNumber = 32000)
peak2 <- get_peak_data(ScanMetadata = TD_ScanMetadata, ScanNumber = 5700, MinAbundance = 0)
peak3 <- get_peak_data(ScanMetadata = TD_ScanMetadata, ScanNumber = 5700, MinAbundance = 2)
peak4 <- get_peak_data(ScanMetadata = RAW_ScanMetadata, ScanNumber = 136, MinAbundance = 0.1)
peak5 <- get_peak_data(ScanMetadata = RAW_ScanMetadata, ScanNumber = 136, MinAbundance = 10)

}

}
