% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_table.R
\name{get_protein_table}
\alias{get_protein_table}
\title{Counts the Proteins in ScanMetadata and generate "Protein Table" object}
\usage{
get_protein_table(
  ScanMetadata,
  FASTAPath,
  QValueMaximum = NULL,
  ScoreMaximum = NULL,
  RemoveContaminants = TRUE
)
}
\arguments{
\item{ScanMetadata}{Object of the scan_metadata class from get_scan_metadata. Required.}

\item{FASTAPath}{File path to the database file, in FASTA format. Required.}

\item{QValueMaximum}{A maximum q-value to filter proteins on. The values range
from 0-1, and NULL means no filter is applied. Default is NULL.}

\item{ScoreMaximum}{A maximum score to filter proteins on. The values range
from 0-1. Remember most scores are small (i.e. 1e-20 or smaller). NULL means no filter is applied.
Default is NULL.}

\item{RemoveContaminants}{A True/False to indicate whether contaminants will be removed
or not. Default is TRUE.}
}
\description{
A simple function which returns a table of protein counts from ScanMetadata
}
\details{
The data.table outputted by this function contains 5 columns of data from the ID file.
\tabular{ll}{
Protein \tab The name of the protein (protein ID) from the ID file \cr
\tab \cr
Number of Peptides \tab The count of peptides identified for that protein. \cr
\tab \cr
Median Q Value \tab The median Q Value for the peptides identified for that protein. \cr
\tab \cr
Median Score \tab The median score for the peptides identified for that protein. \cr
\tab \cr
Description \tab The protein's description as provided in the ID file. \cr
\tab \cr
}

Objects of the class "protein_table" contain attributes of the input variables: QValueMaximum, ScoreMaximum, and RemoveContaminants
}
\examples{
\dontrun{

# Test bottom up data. Note that for these example files, the same FASTA file
# has been used for both bottom-up and top-down
tmpdir <- tempdir()
download.file("https://raw.githubusercontent.com/EMSL-Computing/PSpecteR/master/pspecter_container/TestFiles/QC_Shew.fasta",
              file.path(tmpdir, "QC_Shew.fasta"))

ProteinTable1 <- get_protein_table(BU_ScanMetadata, FASTAPath = file.path(tmpdir, "QC_Shew.fasta"))
ProteinTable2 <- get_protein_table(BU_ScanMetadata, FASTAPath = file.path(tmpdir, "QC_Shew.fasta"), QValueMaximum = 0.1, ScoreMaximum = 0.00003)

}


}
