% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scan_metadata.R
\name{get_scan_metadata}
\alias{get_scan_metadata}
\title{Creates the "Scan Metadata" object from an MS file}
\usage{
get_scan_metadata(MSPath, IDPath = NULL)
}
\arguments{
\item{MSPath}{Data path to the MS file. Currently, mzML, mzXML, and raw are supported.}

\item{IDPath}{Data path to the ID file. Currently, mzID is supported.}
}
\description{
Aligns both MS and ID metadata per scan number. MS data is required. ID data is optional.
}
\details{
The data.table outputted by this function contains 16 main columns of data from both the MS and ID file.
\tabular{ll}{
Scan Number \tab The number of the MS scan where scan 1 was taken at the first retention time. \cr
\tab \cr
MS Level \tab Indiciates whether the scan is a MS1 (precursor) or MS2. \cr
\tab \cr
Retention Time \tab The time, in minutes, that the fragment was eluted. \cr
\tab \cr
Precursor M/Z \tab The M/Z value selected from the MS1 (precursor) scan to fragment for the MS2 scan. Will be NA for MS1 scans. \cr
\tab \cr
Precursor Charge \tab The Z value, charge, from the selected MS1 (precursor) scan. Will be NA for MS1 scans. \cr
\tab \cr
Precursor Scan \tab The MS1 scan number that the MS2 scan was fractioned from. Will be NA for MS1 scans. \cr
\tab \cr
Activation Method \tab A string to indicate which activation (dissociation) method was used to fragment the proteins. \cr
\tab \cr
Sequence \tab The amino acid sequence  (fragment) assigned to the MS2 scan. Only included if IDPath is supplied. \cr
\tab \cr
Protein ID \tab The name of the protein that the amino acid sequence is from. Only included if the IDPath is supplied. \cr
\tab \cr
Calculated Mass \tab The calculated, or expected, mass of the peptide fragment. Only included if the IDPath is supplied. \cr
\tab \cr
Experimental Mass \tab The experimental, or actual, mass that was matched in the MS2 file. Only included if the IDPath is supplied. \cr
\tab \cr
Score \tab A measure of how well the calculated and experimental spectra match. A lower score is better. Only included if the IDPath is supplied. \cr
\tab \cr
Q Value \tab An adjusted p-value that also measures how well the calculated and experimental spectra match. A lower score is preferred. Only included if the IDPath is supplied. \cr
\tab \cr
Decoy \tab A logical (TRUE/FALSE/NA) indicating whether the identified protein is a protein decoy or not. \cr
\tab \cr
Modifications \tab A string indicating whether a modification was identified or not. The format is "Name@Position=MW & Name2@Position2=MW2" etc \cr
\tab \cr
}


Objects of the class "scan_metadata" contain attributes that are referenced by downstream functions.
   The attributes are as follows:
\tabular{ll}{
MSPath \tab The path the user supplied in the MSPath parameter. \cr
\tab \cr
IDPath \tab The path the user supplied in the IDPath parameter. Default is NULL. \cr
\tab \cr
MSFileType \tab A flag to indicate whether the inputted MS file is XML-based (will be read with the mzR package) or RAW (will be read with the rawrr package). \cr
\tab \cr
MSnbase \tab An object from the MSnbase package used to pull XIC data when the MSFileType is XML. If not, then this attribute will be NULL. \cr
\tab \cr
MS1Scans \tab An integer vector listing all the MS1 scans in order. Used by the MS1Plots function. \cr
\tab \cr
}
}
\examples{
\dontrun{

# Create a temporary directory and copy example data there
tmpdir <- tempdir()

files <- c(
 "https://raw.githubusercontent.com/EMSL-Computing/PSpecteR/master/pspecter_container/TestFiles/BottomUp/BottomUp.mzML",
 "https://raw.githubusercontent.com/EMSL-Computing/PSpecteR/master/pspecter_container/TestFiles/BottomUp/BottomUp.mzid",
 "https://raw.githubusercontent.com/EMSL-Computing/PSpecteR/master/pspecter_container/TestFiles/TopDown/TopDown.mzML",
 "https://raw.githubusercontent.com/EMSL-Computing/PSpecteR/master/pspecter_container/TestFiles/TopDown/TopDown.mzid"
)

for (file in files) {
  download.file(file, file.path(tmpdir, tail(unlist(strsplit(file, "/")), 1)))
}

# Test with bottom up data
BU_ScanMetadata <- get_scan_metadata(MSPath = file.path(tmpdir, "BottomUp.mzML"),
                                     IDPath = file.path(tmpdir, "BottomUp.mzid"))

# Test with top down data
TD_ScanMetadata <- get_scan_metadata(MSPath = file.path(tmpdir, "TopDown.mzML"),
                                     IDPath = file.path(tmpdir, "TopDown.mzid"))

# Test with raw top down data
library(rawrr)
rawfile <- file.path(path.package(package = 'rawrr'), 'extdata', 'sample.raw')
RAW_ScanMetadata <- get_scan_metadata(MSPath = rawfile)

}

}
