% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xic.R
\name{get_xic}
\alias{get_xic}
\title{Get the Extracted Ion Chromatogram object "xic"}
\usage{
get_xic(
  ScanMetadata,
  MZ,
  RTRange,
  IsotopeNumber = c(0:5),
  Charges = c(1:3),
  XICTol = 1000,
  Messages = TRUE
)
}
\arguments{
\item{ScanMetadata}{Object of the scan_metadata class from get_scan_metadata. Required.}

\item{MZ}{The mz value to extract the ion chromatogram. Required.}

\item{RTRange}{The retention time range to pull isotope and charge values from. It
should be a vector of length two with the min and max RT in minutes. Required.}

\item{IsotopeNumber}{A vector of isotopes to plot. 1 indicates M+1, etc.
The max isotopes is 5, ranging from 0 to 5 where 0 is the non-isotope peak. Default is 0:5.}

\item{Charges}{A vector indicating which charges to plot. There is a max of 3,
ranging from 1-3. Default is 1:3.}

\item{XICTol}{The tolerance for XIC, used only with raw data.}

\item{Messages}{Should progress messages be printed? Default is TRUE.}
}
\description{
Pulls the intensity and retention times for each isotope and adjacent charge state.
}
\details{
The data.table outputted by this function contains 16 main columns of data from both the MS and ID file.
\tabular{ll}{
Charge \tab The charge (Z) state for each trace of the XIC. \cr
\tab \cr
Isotope \tab Written as "M+n", the n is indicative of each successive isotope after the fragment, where the fragment is M+0 or 0. \cr
\tab \cr
Intensity \tab The intensity at the given m/z and retention time range. \cr
\tab \cr
RT \tab The retention time \cr
\tab \cr
}
}
\examples{
\dontrun{

# Test with bottom up data
xictest1 <- get_xic(ScanMetadata = BU_ScanMetadata, MZ = 659.8596, RTRange = c(58, 60))

# Test with top down raw data
xictest2 <- get_xic(ScanMetadata = RAW_ScanMetadata, MZ = 669.8381, RTRange = c(0, 1))

}

}
