% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ion_bar_plot.R
\name{ion_bar_plot}
\alias{ion_bar_plot}
\title{Make a bar plot of fragments per ion}
\usage{
ion_bar_plot(MatchedPeaks, Interactive = FALSE)
}
\arguments{
\item{MatchedPeaks}{Object of the matched_peaks class from get_matched_peaks. Default is NULL. Required.}

\item{Interactive}{A logical to determine whether the plot should be interactive or not. Default is FALSE.}
}
\description{
Generates a bar plot of counts of all fragments, fragments without
   isotopes, and fragments with unique charge states per ion.
}
\examples{
\dontrun{

# Test with bottom up data
BU_Peak <- get_peak_data(ScanMetadata = BU_ScanMetadata, ScanNumber = 31728)
BU_Match <- get_matched_peaks(ScanMetadata = BU_ScanMetadata, PeakData = BU_Peak)

# Make the plot interactive or not
ion_bar_plot(MatchedPeaks = BU_Match)
ion_bar_plot(MatchedPeaks = BU_Match, Interactive = TRUE)

}

}
