% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molform_functions.R
\name{is_sequence}
\alias{is_sequence}
\title{Ensures a string is an acceptable amino acid sequence}
\usage{
is_sequence(Sequence)
}
\arguments{
\item{Sequence}{An amino acid sequence which should be a single string. Required.}
}
\description{
A simple test function to ensure the provided sequence is an
   acceptable amino acid sequence. ProForma sequences are accepted.
}
\details{
The output will either be a "TRUE" acceptable sequence, or "FALSE"
   unacceptable sequence. An acceptable sequence cannot be NULL,
   will not have any spaces or non-letter characters, will be longer than 1
   amino acid, and will not contain the letters B, J, O, U, X, or Z. ProForma
   notation (i.e. "TES[Acetyl]T") is now accepted.
}
\examples{
\dontrun{

# An acceptable sequence will return TRUE
is_sequence("TEST")
is_sequence("TES[Acetyl]T")
is_sequence("T[58.77]E[Acetyl]ST")

# Unacceptable sequences will return FALSE
is_sequence("TESTSEQUENCE")
is_sequence("TIRED $CIENTIST WORKING L8")
is_sequence("T")
is_sequence("T[UnfinishedEST")

}

}
