% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mass_modified_ion.R
\name{make_mass_modified_ion}
\alias{make_mass_modified_ion}
\title{Creates a modified_ion object for get_matched_peaks}
\usage{
make_mass_modified_ion(Ion, Symbol, AMU_Change)
}
\arguments{
\item{Ion}{A vector of ions to modify. Must be a, b, c, x, y, z. Required.}

\item{Symbol}{An additional symbol to distinguish this ion from the original
a-c or x-z. Acceptable inputs are +, ++, -, --, ^, or ^^. Must be the same
length as the Ion vector. Required.}

\item{AMU_Change}{A vector of the AMU mass changes to apply. Must be the same
length as the Ion vector. Required.}
}
\description{
The modified ion allows for additional ion groups (an a-c or x-z
     ion with a mass shift) to be tested.
}
\details{
Depending on the analysis, users may want to investigate ion groups
   (a-c or x-z) with an additional proton, water loss, etc. This allows for that
   functionality, as well as a user-defined symbol for that change.
}
\examples{
\dontrun{

make_mass_modified_ion(Ion = c("b", "z"),
                       Symbol = c("+", "+"),
                       AMU_Change = c(1.00727647, 1.00727647))

}
}
