% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_molecule.R
\name{make_molecule}
\alias{make_molecule}
\title{Makes a Molecule Dataframe, Helper Function to get_peptide_coverage}
\usage{
make_molecule(AtomList, FormulaOnly = FALSE)
}
\arguments{
\item{AtomList}{A list of atom counts}
}
\description{
This simple function generates molecule data frames from lists for easy addition
}
\details{
A data frame is derived from the list, with each atom count and the formula
}
\examples{
\dontrun{

# Generate a molecule from a list of atoms
make_molecule(AtomList = list("C" = 6, "H" = 12, "O" = 6, "N" = 0))

}

}
