% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_peak_data.R
\name{make_peak_data}
\alias{make_peak_data}
\title{Create a peak_data object from an M/Z and Intensity vector}
\usage{
make_peak_data(MZ, Intensity)
}
\arguments{
\item{MZ}{A vector of MZ values. Positive non-zero values only. Required.}

\item{Intensity}{A vector Intensity values. Positive non-zero values only. Required.}
}
\description{
Generate a peak_data object. The MZ and Intensity vectors must be
    of the same length.
}
\details{
The data.table outputted by this function contains both the M/Z and Intensity vectors of the spectra.
}
\examples{
\dontrun{

make_peak_data("MZ" = c(300, 301, 302), "Intensity" = c(50, 100, 50))

}
}
