% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ptm.R
\name{make_ptm}
\alias{make_ptm}
\title{Creates a modifications_pspecter object for get_matched_peaks}
\usage{
make_ptm(Name, AMU_Change, N_Position, Molecular_Formula = NULL)
}
\arguments{
\item{Name}{A vector of the names of modifications to check. Required.}

\item{AMU_Change}{A vector of the AMU mass changes to apply. Required.}

\item{N_Position}{A vector of vectors of N positions to apply a modification. Required.}

\item{Molecular_Formula}{A list of lists, with atoms written as the name, and the
number of atoms written as the element list(list("C" = 2, "H" = 3, "O" = 1)). Required.}
}
\description{
The modifications objects allows for the generation of a stable
    object for testing post-translational methods (PTMs) as more are discovered.
}
\details{
The modification_object informs PSpecteR where modifications change
   mass values for fragment calculations, as well as in the calculation of expected
   abundances (thus why the molecular formula is needed). Users can choose to not
   include the molecular formula, though isotopic percentages may be slightly off.
}
\examples{
\dontrun{

# Make an acetyl modification on the 2nd residue
PTM <- make_ptm(Name = "Acetyl", AMU_Change = 43.045, N_Position = 2,
                Molecular_Formula = list(list("C" = 2, "H" = 3, "O" = 1)))

}
}
