% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_modifications.R
\name{multiple_modifications}
\alias{multiple_modifications}
\title{A function to generate a list of properly formatted proforma strings}
\usage{
multiple_modifications(Sequence, Modification, AlternativeGlossary = NULL)
}
\arguments{
\item{Sequence}{A valid protein sequence. It may start with "M." which gets removed. Required.}

\item{Modification}{An IsoForma modifications annotation written as 
"PTM,Residue(Positions)[Number of Modifications]". Required.}

\item{AlternativeGlossary}{Try a different glossary. See system.file("extdata", "Unimod_v20220602.csv", package = "pspecterlib)
for formatting.}
}
\description{
A function to generate a list of properly formatted proforma strings
}
\examples{
\dontrun{

# Single modification of a single PTM examples 
multiple_modifications("TRICITIES", "Methyl,I(3)[1]")
multiple_modifications("TRICITIES", "Methyl,I(3,5,7)[1]")

# Multiple modifications of a single PTM example
multiple_modifications("TRICITIES", "Methyl,I(3,5,7)[2]") 

# Multiple modifications with a fixed position
multiple_modifications("TRICITIES", "Methyl,I(3^,5,7)[2]") 

# Multiple modifications with two fixed positions 
multiple_modifications("TRICITIES", "Methyl,I(3^,5,7^)[2]")

# Multiple modifications of a single PTM with any "X" residue
multiple_modifications("TRICITIES", "Methyl,X(1,2,3,4,5,6,7,8,9)[2]")

# Multiple modifications with multiple PTMs examples 
multiple_modifications("TRICITIES", "Methyl,T(1)[1];Acetyl,X(2,4,9)[1]")
multiple_modifications("TRICITIES", "Methyl,T(1,2,3,4,5,6,7,8,9)[1];Acetyl,X(2,4,9)[1]")
multiple_modifications("TRICITIES", "Methyl,T(1^,2,3,4,5,6,7^,8,9)[3];Acetyl,X(2,4,9)[1]") 

}
}
