% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molform_functions.R
\name{multiply_molforms}
\alias{multiply_molforms}
\title{Multiply molecular formula objects}
\usage{
multiply_molforms(molform, scalar)
}
\arguments{
\item{molform}{An object of the as.molform class}

\item{scalar}{A positive integer determining the number of times to multiply 
a molecular formula by.}
}
\description{
Multiply molecular formulas by a constant
}
\examples{
\dontrun{

multiply_molforms(as.molform("C6H12O6"), 3)

}
}
