% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promex_feature_plot.R
\name{promex_feature_plot}
\alias{promex_feature_plot}
\title{Generate a feature map from MS1FT data}
\usage{
promex_feature_plot(MS1FT, Proteins = NULL, Interactive = FALSE)
}
\arguments{
\item{MS1FT}{An ms1ft object generated by get_ms1ft. Required.}

\item{Proteins}{Subset down to specific proteins if Ic_Targets file was provided. Default is NULL.}

\item{Interactive}{A logical to determine whether the plot should be interactive or not. Default is FALSE.}
}
\description{
The feature map visualizes ProMex identified features (Monoiostopic
    Masses per Elution (Retention) Time window) colorized by relative abundance.
    This map is often used to compare one top-down analyses to another, as features
    are assigned to proteins downstream with MSPathFinderT.
}
\examples{
\dontrun{

# Use the ms1ft object outputted from get_ms1ft
promex_feature_plot(MS1FT = MS1FT)
promex_feature_plot(MS1FT = MS1FT, Proteins = c("SO_0239"), Interactive = FALSE)

}
}
