% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_plot.R
\name{sequence_plot}
\alias{sequence_plot}
\title{Plot the Sequence with Matched Fragments Annotations, split by a,b,c and x,y,z annotations}
\usage{
sequence_plot(
  MatchedPeaks,
  IncludeIsotopes = FALSE,
  RemoveChargeAnnotation = FALSE,
  RemoveModification = FALSE,
  WrapLength = 8,
  LabelSize = 3
)
}
\arguments{
\item{MatchedPeaks}{A matched_peaks object generated by get_matched_peaks. Required.}

\item{IncludeIsotopes}{A logical to indicate whether isotopes should be included. Default is FALSE.}

\item{RemoveChargeAnnotation}{A logical to indicate whether the annotation for charge states should be included
in the plot. This does not remove charge states; rather, the annotation for charge states. Default is FALSE.}

\item{RemoveModification}{A logical to indicate whether the modification on the residue
should be removed in the plot. Default is FALSE.}

\item{WrapLength}{An integer to index how many letters should be printed before wrapping.
Default is 10, but the value will vary depending on the size of your visualization.}

\item{LabelSize}{The size of the PTM labels. Default is 3.}
}
\description{
A static plot with the sequence annotated with the lowest ppm error fragment
   per N-terminus (a, b, c) and C-terminus (x, y, z) fragment
}
\examples{
\dontrun{

# Test bottom up data
BU_Peak <- get_peak_data(ScanMetadata = BU_ScanMetadata, ScanNumber = 31728)
BU_Match <- get_matched_peaks(ScanMetadata = BU_ScanMetadata, PeakData = BU_Peak)
BU_Match2 <- get_matched_peaks(ScanMetadata = BU_ScanMetadata, PeakData = BU_Peak, AlternativeSequence = "IGA[Acetyl]VGGTENVSLTQSQMPAHNHLVAASTVSGTVKPLANDIIGAGLNK")

# Make plots
sequence_plot(MatchedPeaks = BU_Match)
sequence_plot(MatchedPeaks = BU_Match, RemoveChargeAnnotation = TRUE)
sequence_plot(MatchedPeaks = BU_Match2)


}

}
