% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xic_plot.R
\name{xic_plot}
\alias{xic_plot}
\title{Make an Extracted Ion Chromatogram}
\usage{
xic_plot(XICobj, Smooth = FALSE, Interactive = FALSE)
}
\arguments{
\item{XICobj}{Object of the xic class from get_xic}

\item{Smooth}{A True/False to indicate whether the plot should be smoothed. Default is FALSE.}

\item{Interactive}{A True/False to indicate whether the plot should be interactive. Default is FALSE.}
}
\description{
Generates an interactive or static plot of a XIC generated by get_xic
}
\examples{
\dontrun{

# Test with bottom up data. See ?get_xic
xic_plot(xictest1)
xic_plot(xictest1, Smooth = TRUE, Interactive = TRUE)

# Test with top down raw data. See ?get_xic
xic_plot(xictest2)

}
}
